// libs/ledger-live-common/src/families/internet_computer/bridge/bridgeHelpers/icpRosetta/types.ts
interface ICPRosettaBalance {
  value: string;
  currency: ICPRosettaCurrency;
}

interface ICPRosettaCurrency {
  symbol: string;
  decimals: number;
}
// ...
export interface ICPRosettaICPRosettaOperation {
  operation_identifier: {
    index: number;
  };
  type: string;
  signature_type?: "secp256k1";
  status?: string;
  account: ICPRosettaAccountIdentifier;
  amount?: ICPRosettaBalance;
  metadata?: {
    operations?: ICPRosettaICPRosettaOperation[];
    full_neurons?: FullNeuron[];
    neuron_infos?: {
      [key: string]: NeuronInfo;
    };
    block_index: number;
    transaction_identifier: ICPRosettaTransactionidentifier;
  };
}

export interface ICPRosettaConstructionPayloadsRequest {
  network_identifier: ICPRosettaNetworkIdentifier;
  operations: ICPRosettaICPRosettaOperation[];
  metadata?: ICPRosettaMetadata;
  public_keys: ICPRosettaPublickey[];
}

export interface ICPRosettaPublickey {
  hex_bytes: string;
  curve_type: string;
}

export interface ICPRosettaMetadata {
  memo?: number;
  created_at?: number;
  ingress_end?: number;
  ingress_start?: number;
}

interface ICPRosettaCurrency {
  symbol: string;
  decimals: number;
}

interface ICPRosettaNetworkIdentifier {
  blockchain: string;
  network: string;
}

export interface ICPRosettaConstructionPayloadsResponse {
  unsigned_transaction: string;
  payloads: {
    account_identifier: ICPRosettaAccountIdentifier;
    hex_bytes: string;
    signature_type: string;
  }[];
}

interface ICPRosettaAccountIdentifier {
  address: string;
}

export interface ICPRosettaConstructionCombineRequest {
  network_identifier: ICPRosettaNetworkIdentifier;
  unsigned_transaction: string; // cbor
  signatures: ICPRosettaSignature[];
}

export interface ICPRosettaSignature {
  signing_payload: {
    account_identifier: ICPRosettaAccountIdentifier;
    hex_bytes: string;
    signature_type: string;
  };
  public_key: ICPRosettaPublickey;
  signature_type: string;
  hex_bytes: string;
}

export interface ICPRosettaConstructionCombineResponse {
  signed_transaction: string;
}

export interface ICPRosettaConstructionSubmitRequest {
  network_identifier: ICPRosettaNetworkIdentifier;
  signed_transaction: string;
}
export interface ICPRosettaConstructionSubmitResponse {
  transaction_identifier: ICPRosettaTransactionidentifier;
  metadata: {
    operations: ICPRosettaICPRosettaOperation[];
  };
}

interface ICPRosettaTransactionidentifier {
  hash: string;
}

export interface ICPRosettaConstructionHashRequest {
  network_identifier: ICPRosettaNetworkIdentifier;
  signed_transaction: string;
}

export interface ICPRosettaConstructionHashResponse {
  transaction_identifier: ICPRosettaTransactionidentifier;
}

export interface ICPRosettaBroadcastResult extends SubmitResponse {
  txnHash?: string;
  blockHeight?: string;
}

export interface ICPRosettaConstructionDeriveRequest {
  network_identifier: ICPRosettaNetworkIdentifier;
  public_key: ICPRosettaPublickey;
}

export interface ICPRosettaConstructionDeriveResponse {
  account_identifier: ICPRosettaAccountIdentifier;
}